pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-bioresearch"
version = "0.2.0"
description = "Multi-agent GRPO OpenEnv for biomedical reasoning — 5 tasks including a multi-turn Virtual Tumor Board with specialist actors and tool use"
requires-python = ">=3.10"
dependencies = [
    "openenv-core[core]>=0.2.2,<0.3",
    "openai>=1.40.0",
    "pydantic>=2.5,<3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0,<10.0",
    "pytest-cov>=4.0.0,<9.0",
    "pytest-asyncio>=0.23.0,<2.0",
]
playground = [
    "gradio>=6.0.0,<7.0",
]
train = [
    "trl>=0.14.0",
    "transformers>=4.45.0",
    "accelerate>=0.34.0",
    "peft>=0.12.0",
    "datasets>=2.20.0",
    "matplotlib>=3.8.0",
]

[project.scripts]
server = "bioresearch.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["bioresearch", "bioresearch.server"]
package-dir = { "bioresearch" = ".", "bioresearch.server" = "server" }
